Revert "Feature login"
