+++ index.html
@@ -2,11 +2,9 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700;800;900&display=swap" rel="stylesheet">
-    <link rel="icon" type="image/x-icon" href="/src/assets/images/icon-favicon.svg">
-    <meta name="referrer" content="no-referrer" />
-    <title>MotoLink</title>
+    <title>Vite + React + TS</title>
   </head>
   <body>
     <div id="root"></div>
