+++ src/App.tsx
@@ -1,14 +1,13 @@
-import { RounterMain } from "./Routes/RouterMain";
-import { GlobalStyles } from "./styles/global";
-import { ToastContainer } from "react-toastify";
-import "react-toastify/dist/ReactToastify.css";
+import { RounterMain } from './assets/Routes/RouterMain'
 
-export const App = () => {
+function App() {
+
+
   return (
-    <div className="App">
-      <ToastContainer />
-      <RounterMain />
-      <GlobalStyles />
-    </div>
-  );
-};
+   <>
+   <RounterMain/>
+   </>
+  )
+}
+
+export default App
