+++ src/assets/Pages/DashBoardEmpresa/DashBoardEmprea.tsx
@@ -1,0 +1,6 @@
+export const DashEmpresas=()=>{
+
+    return(
+        <h1>Dash Empresas</h1>
+    )
+}
