+++ src/assets/Pages/DashboardMotoboy/DashBoardMotoboy.tsx
@@ -1,0 +1,6 @@
+export const DashMotoboy=()=>{
+
+    return(
+        <h1>Dash Motoboy</h1>
+    )
+}
