+++ src/assets/Pages/Login/Login.tsx
@@ -1,0 +1,6 @@
+export const Login=()=>{
+
+    return(
+        <h1>Login</h1>
+    )
+}
