+++ src/assets/Pages/Register/Register.tsx
@@ -1,0 +1,6 @@
+export const Register=()=>{
+
+    return(
+        <h1>Register</h1>
+    )
+}
