--- src/components/LoginForm/loginSchema.tsx
@@ -1,7 +1,0 @@
-import { z } from "zod";
-
-export const LoginSchema = z.object({
-  email: z.string().nonempty("O email é obrigatório"),
-
-  password: z.string().nonempty("A Senha é obrigatória"),
-});
