--- src/components/LoginForm/style.ts
@@ -1,20 +1,0 @@
-import styled from "styled-components";
-
-export const StyledForm = styled.div`
-  display: flex;
-  gap: 20px;
-  flex-direction: column;
-  justify-content: center;
-
-  h3 {
-    align-self: center;
-  }
-
-  form {
-    width: 100%;
-    display: flex;
-    align-self: center;
-    gap: 20px;
-    flex-direction: column;
-  }
-`;
